dont return active snaphost during previous request, log timestamp
while testing with go bindings for the proxmox-backup-qemu api, i
noticed that on a backup run for a fixed image backup the following
warning is tracked during each backup attempt:

 [WARNING] File test.img.fidx not found in snapshot backups/testbackup|1725023177|vm

via direct api call, it is possible to upload an config file blob before
the actual image is created. That means once the request to create the
image is done, the snapshot directory already exists.

So if the client backups an blob first, and then creates an image
backup within one backup session, the previous endpoint returns
the snapshot that is actually beeing written, because the endpoint
is called seperately for each file:

 New Backup request from host [127.0.0.1:41706], ID: [testbackup], Type: [vm], Bucket: [backups] Timestamp: [1725023177]
 [..]
 [  INFO ] Latest snapshot for backup id [testbackup]: [backups/testbackup|1725023088|vm] < previous one
 [  INFO ] Handling get request for previous (test.img.fidx)
 [  INFO ] Latest snapshot for backup id [testbackup]: [backups/testbackup|1725023177|vm] < the actual one
 [WARNING] File test.img.fidx not found in snapshot backups/testbackup|1725023177|vm ([{test.blob none %!s(uint64=17)}])

GetLatestSnapshot now is passed backup id and backup time and
if the ListSnapshots function returns the current actively
processed Snapshot, it skips.